requirements.txt
Django==2.1.2
django-render-block==0.5
django-templated-email==2.2.0
pkg-resources==0.0.0
pytz==2018.7
six==1.11.0
